+++ src/api/routes/admin/orders-packer.js
@@ -1,0 +1,6 @@
+export default async (req, res) => {
+  const shippoFulfillmentService = req.scope.resolve("shippoFulfillmentService")
+  const { order_id } = req.params
+
+  res.json(await shippoFulfillmentService.retrievePackerResults(order_id))
+}